Move File.deleteRecursivelyWithoutSymlinks to FileUtils
