+++ core/src/main/kotlin/io/paddle/plugin/standard/tasks/CleanTask.kt
@@ -3,9 +3,9 @@ package io.paddle.plugin.standard.tasks
 import io.paddle.project.Project
 import io.paddle.project.Tasks
 import io.paddle.tasks.Task
+import io.paddle.utils.deleteRecursivelyWithoutSymlinks
 import io.paddle.utils.tasks.TaskDefaultGroups
 import java.io.File
-import java.nio.file.Files
 
 class CleanTask(project: Project) : Task(project) {
     var locations = ArrayList<File>()
@@ -27,11 +27,3 @@ class CleanTask(project: Project) : Task(project) {
 
 val Tasks.clean: CleanTask
     get() = this.getOrFail("clean") as CleanTask
-
-fun File.deleteRecursivelyWithoutSymlinks() {
-    if (this.isDirectory && !Files.isSymbolicLink(this.toPath())) {
-        val children = this.listFiles() ?: return
-        children.forEach { it.deleteRecursivelyWithoutSymlinks() }
-    }
-    this.delete()
-}
